+ Added "Elements to be Integrated" section with a short description of what insert.  + Review and extended the "Entry criteria". Almost done.
About the "Elements to be Integrated" section: here described (well, please) the subsystems composing the final software system. The identification
of these subsystems MUST follows the system's design (the one in the Design Document). To have hints, follow the ITPD of De Nicolao and the
following one:
		https://github.com/valerio-castelli/casaticastelli_sweng2/blob/master/Deliveries/Final%20release/4_ITPD_rev11.pdf

(The last one is the one shown in the last lecture [that you missed] of SE2).

NB: I don't think we should talk about the strategy adopted to integrate the subsystems. There's a dedicated section (2.3 section of ITPD) for
the integration testing strategy.

Cheers.